+++ src/core/size.js
@@ -5,8 +5,8 @@ THREE.Bootstrap.registerPlugin('size', {
     height: null,
     aspect: null,
     scale: 1,
-    capWidth: Infinity,
-    capHeight: Infinity,
+    maxRenderWidth: Infinity,
+    maxRenderHeight: Infinity,
   },
 
   listen: [
@@ -74,9 +74,9 @@ THREE.Bootstrap.registerPlugin('size', {
       aspect = w / h;
     }
 
-    // Apply scale and resolution cap
-    rw = Math.min(w * options.scale, options.capWidth);
-    rh = Math.min(h * options.scale, options.capHeight);
+    // Apply scale and resolution max
+    rw = Math.min(w * options.scale, options.maxRenderWidth);
+    rh = Math.min(h * options.scale, options.maxRenderHeight);
 
     // Retain aspect ratio
     raspect = rw / rh;
@@ -87,25 +87,12 @@ THREE.Bootstrap.registerPlugin('size', {
       rh = Math.round(rw / aspect);
     }
 
-    if (renderer) {
-      var el = renderer.domElement;
-
-      // Resize renderer to render width if it's a canvas
-      if (el && el.tagName == 'CANVAS') {
-        renderer.setSize(rw, rh);
-      }
-      // Or real width if it's just a DOM element
-      else {
-        renderer.setSize(w, h);
-      }
-
-      // Resize Canvas
-      style = renderer.domElement.style;
-      style.width = w + "px";
-      style.height = h + "px";
-      style.marginLeft = ml + "px";
-      style.marginTop = mt + "px";
-    }
+    // Resize and position renderer element
+    style = renderer.domElement.style;
+    style.width = w + "px";
+    style.height = h + "px";
+    style.marginLeft = ml + "px";
+    style.marginTop = mt + "px";
 
     // Notify
     _.extend(three.Size, {