+++ test/core/renderer.spec.js
@@ -53,4 +53,56 @@ describe('renderer', function () {
     document.body.removeChild(element);
   });
 
+  it("calls renderer setSize", function () {
+
+    var element = document.createElement('div');
+    document.body.appendChild(element);
+
+    var options = {
+      init: false,
+      element: element,
+      plugins: ['renderer'],
+    };
+
+    var three = new THREE.Bootstrap(options);
+
+    var called = 0;
+    var callback = function () { called++; };
+
+    three.init();
+    three.renderer.setSize = callback;
+    three.plugins.renderer.resize({ renderWidth: 5, renderHeight: 4, viewWidth: 3, viewHeight: 2, aspect: 3/2 }, three);
+    three.destroy();
+
+    expect(called).toBe(1);
+  });
+
+  it("calls renderer setSize and setRenderSize", function () {
+
+    var element = document.createElement('div');
+    document.body.appendChild(element);
+
+    var options = {
+      init: false,
+      element: element,
+      plugins: ['renderer'],
+    };
+
+    var three = new THREE.Bootstrap(options);
+
+    var called = 0;
+    var callback = function () { called++; };
+
+    three.init();
+    el = three.renderer.domElement
+    three.renderer.domElement = document.createElement('div')
+    three.renderer.setSize = callback;
+    three.renderer.setRenderSize = callback;
+    three.plugins.renderer.resize({ renderWidth: 5, renderHeight: 4, viewWidth: 3, viewHeight: 2, aspect: 3/2 }, three);
+    three.renderer.domElement = el
+    three.destroy();
+
+    expect(called).toBe(2);
+  });
+
 });